+++ packages/cfd/src/Constants/cfd_compare_account_content.ts
@@ -261,12 +261,3 @@ export const preppstore_eu_demo_content: TCompareAccountContentProps[] = [
         },
     },
 ];
-
-export const ACCOUNT_FLAG = Object.freeze({
-    CR_DEMO_CONTENT: 'show_cr_demo_account',
-    LOW_RISK_CR_EU_CONTENT: 'show_low_risk_cr_eu',
-    LOW_RISK_CR_NON_EU_CONTENT: 'show_low_risk_cr_non_eu',
-    EU_DEMO_CONTENT: 'show_eu_demo',
-    EU_REAL_CONTENT: 'show_eu_real',
-    HIGH_RISK_CR: 'show_high_risk_cr',
-});
