+++ packages/cfd/src/Containers/props.types.ts
@@ -289,38 +289,36 @@ export type TCompareAccountContentProps = {
 export type TCompareAccountFooterButtonData = { label: string; action: string };
 
 export type TDMT5CompareModalContentProps = {
-    context: RootStore;
+    account_flag: string;
     account_settings: GetSettings;
-    setAccountSettings: (get_settings_response: GetSettings) => void;
+    account_status: GetAccountStatus;
     account_type: TOpenAccountTransferMeta;
-    setAccountType: (account_type: TOpenAccountTransferMeta) => void;
     clearCFDError: () => void;
+    context: RootStore;
     current_list: Record<string, DetailsOfEachMT5Loginid>;
     has_real_account: boolean;
-    is_logged_in: boolean;
     is_demo_tab: boolean;
+    is_logged_in: boolean;
+    is_pre_appstore_setting: boolean;
+    is_preappstore_cr_demo_account: boolean;
     is_real_enabled: boolean;
     is_virtual: boolean;
     openDerivRealAccountNeededModal: () => void;
     openPasswordModal: (account_type: TOpenAccountTransferMeta) => void;
     openSwitchToRealAccountModal: () => void;
-    toggleCompareAccounts: () => void;
-    toggleCFDVerificationModal: () => void;
-    trading_platform_available_accounts: TTradingPlatformAvailableAccount[];
-    toggleCFDPersonalDetailsModal: (is_from_mt5_compare_accounts?: boolean) => void;
-    setJurisdictionSelectedShortcode: (shortcode: string) => void;
-    is_eu_client: boolean;
-    account_status: GetAccountStatus;
-    upgradeable_landing_companies: unknown[];
-    setAppstorePlatform: (platform: string) => void;
-    should_show_derivx: boolean;
-    should_restrict_bvi_account_creation: boolean;
-    updateAccountStatus: () => void;
     real_account_creation_unlock_date: string;
+    setAccountSettings: (get_settings_response: GetSettings) => void;
+    setAccountType: (account_type: TOpenAccountTransferMeta) => void;
+    setAppstorePlatform: (platform: string) => void;
+    setJurisdictionSelectedShortcode: (shortcode: string) => void;
     setShouldShowCooldownModal: (value: boolean) => void;
-    show_preappstore_eu_demo: boolean;
-    is_preappstore_cr_demo_account: boolean;
+    should_restrict_bvi_account_creation: boolean;
+    should_show_derivx: boolean;
     show_eu_related: boolean;
-    is_pre_appstore_setting: boolean;
-    account_flag: string;
+    toggleCFDPersonalDetailsModal: (is_from_mt5_compare_accounts?: boolean) => void;
+    toggleCFDVerificationModal: () => void;
+    toggleCompareAccounts: () => void;
+    trading_platform_available_accounts: TTradingPlatformAvailableAccount[];
+    updateAccountStatus: () => void;
+    upgradeable_landing_companies: unknown[];
 };
