+++ packages/shared/src/utils/constants/index.ts
@@ -1,2 +1,3 @@
 export * from './barriers';
 export * from './contract';
+export * from './account_flags';
